+++ lua/obsidian/init.lua
@@ -148,6 +148,17 @@ obsidian.setup = function(opts)
         require("obsidian.completion.plugin_initializers.blink").inject_sources()
       end
 
+      local win = vim.api.nvim_get_current_win()
+
+      vim.treesitter.start(ev.buf, "markdown") -- for when user don't use nvim-treesitter
+
+      vim.wo[win].foldmethod = "expr"
+      vim.wo[win].foldexpr = "v:lua.vim.treesitter.foldexpr()"
+      vim.wo[win].fillchars = "foldopen:,foldclose:,fold: ,foldsep: "
+      vim.wo[win].foldtext = ""
+      vim.wo[win].foldlevel = 99
+      vim.wo[win].smoothscroll = true
+
       -- Run enter-note callback.
       client.callback_manager:enter_note(function()
         return obsidian.Note.from_buffer(ev.bufnr)